fix: separate client component to fix params typing issue on post-test detail
